[Refactor] Move api options to its own class (#6512)
