+++ besu/src/test/java/org/hyperledger/besu/cli/BesuCommandTest.java
@@ -1040,109 +1040,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcMaxLogsRangeOptionMustBeUsed() {
-
-    final long rpcMaxLogsRange = 150L;
-    parseCommand("--rpc-max-logs-range", Long.toString(rpcMaxLogsRange));
-
-    verify(mockRunnerBuilder).apiConfiguration(apiConfigurationCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(apiConfigurationCaptor.getValue())
-        .isEqualTo(ImmutableApiConfiguration.builder().maxLogsRange((rpcMaxLogsRange)).build());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcGasCapOptionMustBeUsed() {
-    final long rpcGasCap = 150L;
-    parseCommand("--rpc-gas-cap", Long.toString(rpcGasCap));
-
-    verify(mockRunnerBuilder).apiConfiguration(apiConfigurationCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(apiConfigurationCaptor.getValue())
-        .isEqualTo(ImmutableApiConfiguration.builder().gasCap((rpcGasCap)).build());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void apiPriorityFeeLimitingEnabledOptionMustBeUsed() {
-    parseCommand("--api-gas-and-priority-fee-limiting-enabled");
-    verify(mockRunnerBuilder).apiConfiguration(apiConfigurationCaptor.capture());
-    verify(mockRunnerBuilder).build();
-    assertThat(apiConfigurationCaptor.getValue())
-        .isEqualTo(
-            ImmutableApiConfiguration.builder().isGasAndPriorityFeeLimitingEnabled(true).build());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void apiPriorityFeeLowerBoundCoefficientOptionMustBeUsed() {
-    final long lowerBound = 150L;
-    parseCommand(
-        "--api-gas-and-priority-fee-lower-bound-coefficient",
-        Long.toString(lowerBound),
-        "--api-gas-and-priority-fee-limiting-enabled");
-    verify(mockRunnerBuilder).apiConfiguration(apiConfigurationCaptor.capture());
-    verify(mockRunnerBuilder).build();
-    assertThat(apiConfigurationCaptor.getValue())
-        .isEqualTo(
-            ImmutableApiConfiguration.builder()
-                .lowerBoundGasAndPriorityFeeCoefficient(lowerBound)
-                .isGasAndPriorityFeeLimitingEnabled(true)
-                .build());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void
-      apiPriorityFeeLowerBoundCoefficients_MustNotBeGreaterThan_apiPriorityFeeUpperBoundCoefficient() {
-    final long lowerBound = 200L;
-    final long upperBound = 100L;
-
-    parseCommand(
-        "--api-gas-and-priority-fee-limiting-enabled",
-        "--api-gas-and-priority-fee-lower-bound-coefficient",
-        Long.toString(lowerBound),
-        "--api-gas-and-priority-fee-upper-bound-coefficient",
-        Long.toString(upperBound));
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "--api-gas-and-priority-fee-lower-bound-coefficient cannot be greater than the value of --api-gas-and-priority-fee-upper-bound-coefficient");
-  }
-
-  @Test
-  public void apiPriorityFeeUpperBoundCoefficientsOptionMustBeUsed() {
-    final long upperBound = 200L;
-    parseCommand(
-        "--api-gas-and-priority-fee-upper-bound-coefficient",
-        Long.toString(upperBound),
-        "--api-gas-and-priority-fee-limiting-enabled");
-    verify(mockRunnerBuilder).apiConfiguration(apiConfigurationCaptor.capture());
-    verify(mockRunnerBuilder).build();
-    assertThat(apiConfigurationCaptor.getValue())
-        .isEqualTo(
-            ImmutableApiConfiguration.builder()
-                .upperBoundGasAndPriorityFeeCoefficient(upperBound)
-                .isGasAndPriorityFeeLimitingEnabled(true)
-                .build());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
   public void p2pPeerUpperBound_without_p2pPeerLowerBound_shouldSetLowerBoundEqualToUpperBound() {
 
     final int maxPeers = 23;
